fix: Dir

+++ Source/Server/Game/Network/NetworkSend.cs
@@ -769,7 +769,7 @@ namespace Server
                 buffer.WriteInt32((int)Data.MapNpc[mapNum].Npc[i].Num);
                 buffer.WriteInt32(Data.MapNpc[mapNum].Npc[i].X);
                 buffer.WriteInt32(Data.MapNpc[mapNum].Npc[i].Y);
-                buffer.WriteInt32(Data.MapNpc[mapNum].Npc[i].Dir);
+                buffer.WriteByte(Data.MapNpc[mapNum].Npc[i].Dir);
                 for (int x = 0; x < vitalCount; x++)
                 {
                     buffer.WriteInt32(Data.MapNpc[mapNum].Npc[i].Vital[x]);
